requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pika==1.3.2
sqlalchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
selenium==4.15.0
fake-useragent==1.4.0
